refactor: Responses 분리 및 BaseResponse 정리
* `Responses`를 별도 파일로 분리하여 코드 가독성 향상
* `BaseResponse`에 불필요한 `Responses` 관련 코드 제거
