+++ src/main/kotlin/com/wq/demo/web/common/BaseResponse.kt
@@ -1,0 +1,8 @@
+package com.wq.demo.web.common
+
+data class BaseResponse<T> (
+    val success: Boolean,
+    val message: String,
+    val data: T?,
+    val error: String? = null
+)
