+++ src/main/kotlin/com/wq/demo/web/common/Responses.kt
@@ -2,13 +2,6 @@ package com.wq.demo.web.common
 
 import com.wq.demo.shared.error.ApiCode
 
-data class BaseResponse<T> (
-    val success: Boolean,
-    val message: String,
-    val data: T?,
-    val error: String? = null
-)
-
 object Responses {
     fun <T> success(
         message: String = "요청에 성공적으로 응답하였습니다.",
